+++ sharding-jdbc-spring/sharding-jdbc-spring-namespace/src/main/java/io/shardingsphere/jdbc/spring/datasource/SpringMasterSlaveDataSource.java
@@ -25,7 +25,9 @@ import io.shardingsphere.core.jdbc.core.datasource.MasterSlaveDataSource;
 import javax.sql.DataSource;
 import java.sql.SQLException;
 import java.util.Collection;
+import java.util.LinkedHashMap;
 import java.util.Map;
+import java.util.Properties;
 
 /**
  * Master-slave datasource for spring namespace.
@@ -34,14 +36,17 @@ import java.util.Map;
  */
 public class SpringMasterSlaveDataSource extends MasterSlaveDataSource {
     
-    public SpringMasterSlaveDataSource(final Map<String, DataSource> dataSourceMap, final String name, final String masterDataSourceName, final Collection<String> slaveDataSourceNames, 
-                                       final MasterSlaveLoadBalanceAlgorithm strategy, final Map<String, Object> configMap) throws SQLException {
-        super(dataSourceMap, getMasterSlaveRuleConfiguration(name, masterDataSourceName, slaveDataSourceNames, strategy), configMap);
+    public SpringMasterSlaveDataSource(final Map<String, DataSource> dataSourceMap, final String name,
+                                       final String masterDataSourceName, final Collection<String> slaveDataSourceNames, final MasterSlaveLoadBalanceAlgorithm strategy, final Map<String, Object> configMap, final Properties props) throws SQLException {
+        super(dataSourceMap, getMasterSlaveRuleConfiguration(name, masterDataSourceName, slaveDataSourceNames, strategy),
+                null == configMap ? new LinkedHashMap<String, Object>() : configMap, null == props ? new Properties() : props);
     }
     
-    public SpringMasterSlaveDataSource(final Map<String, DataSource> dataSourceMap, final String name, final String masterDataSourceName, final Collection<String> slaveDataSourceNames, 
-                                       final MasterSlaveLoadBalanceAlgorithmType strategyType, final Map<String, Object> configMap) throws SQLException {
-        super(dataSourceMap, getMasterSlaveRuleConfiguration(name, masterDataSourceName, slaveDataSourceNames, null == strategyType ? null : strategyType.getAlgorithm()), configMap);
+    public SpringMasterSlaveDataSource(final Map<String, DataSource> dataSourceMap, final String name,
+                                       final String masterDataSourceName, final Collection<String> slaveDataSourceNames, final MasterSlaveLoadBalanceAlgorithmType strategyType, final Map<String, Object> configMap, final Properties props) throws SQLException {
+        super(dataSourceMap,
+                getMasterSlaveRuleConfiguration(name, masterDataSourceName, slaveDataSourceNames, null == strategyType ? null : strategyType.getAlgorithm()),
+                null == configMap ? new LinkedHashMap<String, Object>() : configMap, null == props ? new Properties() : props);
     }
     
     private static MasterSlaveRuleConfiguration getMasterSlaveRuleConfiguration(